+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Platform setup script that is ran with `python -m acconeer.exptool.setup`.
+- Sensor selection for Presence- and Distance detector.
 
 ## v5.1.0